Optimize.
Fix computation of linear distance along the front chain. We can then optimize
by reusing the distance computed during traversal.

Note: the computed linear distance is under-estimated when there are gaps (non-
tangent adjacent circles) in the front chain. This state may occur temporarily
before a circle is placed, but it doesn’t seem to interfere with the algorithm.