New place() function that avoids overlaps
The old place(a, b, c) function would sometimes place c so it overlapped
with a and/or b, if c had a much larger radius than a and b.

This version doesn’t seem to.

As per discussion on #111.

This commit also updates the tests in find-place-bugs.js to
use the new function, and updates the intersection test in
siblings-test.js to match

Note that flare-test.js fails. I haven’t investigated why: I hope it is simply
a result of the changes in numerical precision causing slightly different
circle placements.